Updated docs, improved examples
