+++ docs/api.rst
@@ -1,8 +1,19 @@
 
-.. If you created a package, create one automodule per module in the package.
 
-.. If your library file(s) are nested in a directory (e.g. /adafruit_foo/foo.py)
-.. use this format as the module name: "adafruit_foo.foo"
+.. automodule:: adafruit_magtag.fakerequests
+   :members:
 
-.. automodule:: adafruit_magtag
+.. automodule:: adafruit_magtag.graphics
+   :members:
+
+.. automodule:: adafruit_magtag.magtag
+   :members:
+
+.. automodule:: adafruit_magtag.network
+   :members:
+
+.. automodule:: adafruit_magtag.wifi_module
+   :members:
+
+.. automodule:: adafruit_magtag.peripherals
    :members:
