+++ docs/examples.rst
@@ -6,3 +6,10 @@ Ensure your device works with this simple test.
 .. literalinclude:: ../examples/magtag_simpletest.py
     :caption: examples/magtag_simpletest.py
     :linenos:
+
+Other Demos
+------------
+
+.. literalinclude:: ../examples/bitcoin_demo.py
+    :caption: examples/bitcoin_demo.py
+    :linenos:
