+++ examples/magtag_simpletest.py
@@ -1,3 +1,36 @@
 # SPDX-FileCopyrightText: 2017 Scott Shawcroft, written for Adafruit Industries
 #
 # SPDX-License-Identifier: Unlicense
+import time
+import terminalio
+from adafruit_magtag.magtag import MagTag
+
+magtag = MagTag()
+
+magtag.add_text(
+    text_font=terminalio.FONT,
+    text_position=(
+        50,
+        (magtag.graphics.display.height // 2) - 1,
+    ),
+    text_scale=3,
+)
+
+magtag.set_text("Hello World")
+
+buttons = magtag.peripherals.buttons
+button_colors = ((255, 0, 0), (255, 150, 0), (0, 255, 255), (180, 0, 255))
+button_tones = (1047, 1318, 1568, 2093)
+timestamp = time.monotonic()
+
+while True:
+    for i, b in enumerate(buttons):
+        if not b.value:
+            print("Button %c pressed" % chr((ord("A") + i)))
+            magtag.peripherals.neopixel_disable = False
+            magtag.peripherals.neopixels.fill(button_colors[i])
+            magtag.peripherals.play_tone(button_tones[i], 0.25)
+            break
+    else:
+        magtag.peripherals.neopixel_disable = True
+    time.sleep(0.01)
